backend/requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
playwright
groq
python-dotenv
